got build scripts working without compile errors, still runtime errors
