+++ app/app.html
@@ -1,7 +1,7 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <link rel="stylesheet" href="./node_modules/bootstrap/dist/css/bootstrap.min.css">
+    <link rel="stylesheet" href="../node_modules/bootstrap/dist/css/bootstrap.min.css">
     <!-- <link rel="stylesheet" href="./node_modules/datatables.net-bs/css/dataTables.bootstrap.css"> -->
     <link rel="stylesheet" href="./css/pokenurse.css">
     <title>PokéNurse &bull; Login</title>
@@ -9,8 +9,8 @@
   <body>
     <div id="content"></div>
     <script>
-      require('babel-register')
-      require("./js") // Renders App into id="content"
+      // require('babel-register')
+      require("../dist/bundle.js") // Renders App into id="content"
     </script>
   </body>
 </html>
