+++ app/screens/Detail/index.jsx
