+++ app/screens/Table/components/Pokemon.jsx
