+++ generateBaseStats/GAME_MASTER_v0_1.json
