+++ generateBaseStats/GAME_MASTER_v0_2.json
