+++ generateBaseStats/basicAttacks.json
