+++ generateBaseStats/chargedAttacks.json
