+++ generateBaseStats/cpStats.json
