+++ generateBaseStats/index.js
