+++ webpack.config.base.js
@@ -29,5 +29,9 @@ export default {
   externals: [
     // put your node 3rd party libraries which can't be built with webpack here
     // (mysql, mongodb, and so on..)
+    'pogobuf',
+    'node-pogo-protos',
+    'node-pogo-signature',
+    'electron-devtools-installer'
   ]
 };
